replace int by SOCKET

+++ tws_api/EClientSocket.h
@@ -48,7 +48,7 @@ private:
 	void encodeMsgLen(std::string& msg, unsigned offset) const;
 public:
 	bool handleSocketError();
-	int receive( char* buf, int sz);
+	int receive( char* buf, size_t sz);
 
 public:
 	// callback from socket
